requirements.txt
streamlit
transformers
torch
gTTS
requests
Pillow